requirements.txt
comtypes==1.4.6
photoshop-python-api==0.22.4
